requirements.txt
asgiref==3.8.1
attrs==25.3.0
Brotli==1.1.0
cffi==1.17.1
cssselect2==0.8.0
dj-database-url==3.0.0
Django==5.2.3
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.6.1
fonttools==4.58.4
gunicorn==22.0.0
inflection==0.5.1
Jinja2==3.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
pyphen==0.17.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.14.0
uritemplate==4.2.0
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1
